+++ library/core/designsystem/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/core/designsystem/ui/icons/ToolkitIconContent.kt
@@ -54,6 +54,13 @@ fun ToolkitIconContent(
     tint: Color = LocalContentColor.current,
 ) {
     when (icon) {
+        is ToolkitIcon.Lottie -> LottieToolkitIcon(
+            icon = icon,
+            contentDescription = contentDescription,
+            modifier = modifier,
+            atEnd = atEnd,
+            tint = tint,
+        )
         is ToolkitIcon.Vector -> {
             Icon(
                 imageVector = icon.imageVector,